ISSUE #4736 - separate markupStage from screenshotDialog
