+++ frontend/src/v4/routes/components/screenshotDialog/components/typingHandler/typingHandler.component.tsx
@@ -18,7 +18,7 @@ import { useState, useEffect, useCallback, CSSProperties } from 'react';
 import Konva from 'konva';
 import { isEmpty } from 'lodash';
 
-import { MODES } from '../../screenshotDialog.helpers';
+import { MODES } from '../../markupStage/markupStage.helpers';
 import { EditableText } from '../editableText/editableText.component';
 
 interface IProps {
