+++ frontend/src/v4/routes/components/screenshotDialog/screenshotDialog.container.ts
@@ -18,32 +18,23 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { createStructuredSelector } from 'reselect';
-import { withViewer } from '../../../services/viewer/viewer';
 
 import {
 	selectAreFutureElements,
 	selectArePastElements,
-	selectCanvasElements,
 	CanvasHistoryActions
 } from '../../../modules/canvasHistory';
-import { selectPathname } from '../../../modules/router/router.selectors';
 import { ScreenshotDialog } from './screenshotDialog.component';
 
 const mapStateToProps = createStructuredSelector({
-	canvasElements: selectCanvasElements,
 	arePastElements: selectArePastElements,
 	areFutureElements: selectAreFutureElements,
-	pathname: selectPathname,
 });
 
 export const mapDispatchToProps = (dispatch) => bindActionCreators({
-	addElement: CanvasHistoryActions.add,
 	updateElement: CanvasHistoryActions.update,
-	removeElement: CanvasHistoryActions.remove,
 	undo: CanvasHistoryActions.undo,
 	redo: CanvasHistoryActions.redo,
-	clearHistory: CanvasHistoryActions.clearHistory,
-	initHistory: CanvasHistoryActions.initHistory
 }, dispatch);
 
-export default withViewer(connect(mapStateToProps, mapDispatchToProps)(ScreenshotDialog));
+export default connect(mapStateToProps, mapDispatchToProps)(ScreenshotDialog);
